EnemyManager class creation
