--- Assets/Code/Scripts/Enemies/SquadSpawner.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d719866dce005bc47b09ce95e8f09292
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
